Tiles used for blocks now have their own file: block_tiles.bmp There are 1024 tiles in a 32x32 grid, just like the regular tiles.bmp They are loaded into btiles[1024] just like the other tiles use tiles[1024]. All tiles used to draw blocks now use btiles[] instead of tiles[]
Completely redid the block attribute method.
The old method used the tile number as an indicator of the block properties.
The new method uses bit flags stored in the same int as the tile number

Made some nice graphical tools to manipulate tiles and .bmp file and flags.

Added a new page to the html documentation 'block_flags.html'

Signed-off-by: mweiss001 <mweiss001@gmail.com>